pyproject.toml
[tool.poetry]
name = "nonebot-plugin-couplets"
version = "0.2.0"
description = "Nonebot2对对联插件。"
authors = ["CMHopeSunshine <277073121@qq.com>"]
readme = "README.md"
license = "AGPL"
keywords = ["nonebot", "nonebot2", "couplets"]
homepage = "https://github.com/CMHopeSunshine/nonebot-plugin-couplets"

[[tool.poetry.source]]
name = "ali"
default = true
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
